+++ app/src/main/java/id/my/bananapixel/quakealert/db/QuakeRepository.kt
@@ -33,7 +33,7 @@ class QuakeRepository(
                         latitude = report.latitude.let { if (it.isNaN()) 0.0 else it },
                         longitude = report.longitude.let { if (it.isNaN()) 0.0 else it },
                         pga = report.pga_maks.ifEmpty { "0" },
-                        durasi = report.durasi,
+                        durasi = report.durasi.toInt(),
                         station_id = report.station_id.ifEmpty { "N/A" },
                         intensity = report.intensitas_maks.ifEmpty { "I" }
                     )
